// TPViewProject/src/app/models/TPDocumentation.ts
export interface TPDocumentation {
    id: number;
    identificationTypeId: number;
    identificationNumber: number;
    companyName: string;
    firstName: string;
    secondName: string;
    firstLastName: string;
    secondLastName: string;
    email: string;
    checkMessagesCell: boolean;
    checkMessagesEmail: boolean;
}